CheckPrimeNumber/src/com/anhpd/checkprime/core/CheckPrime.java
package com.anhpd.checkprime.core;

/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */

/**
 *
 * @author ADMIN
 */
public class CheckPrime {
    public static boolean isPrime(int num) {
        if (num <= 1) {
            return false;
        }

        for (int i = 2; i <= num / 2; i++) {
            if (num % i == 0) {
                return true;  // Bug: should be false
            }
        }

        return true;
    }
}
